+++ terra_geocrud/serializers.py
@@ -203,6 +203,14 @@ class CrudFeatureListSerializer(BaseUpdatableMixin, FeatureSerializer):
         return reverse('feature-detail',
                        args=(obj.layer_id, obj.identifier))
 
+    def get_relations(self, obj):
+        return {
+            relation.name: reverse('feature-relation',
+                                   args=(obj.layer_id, obj.identifier, relation.pk))
+            for relation in obj.layer.relations_as_origin.all()
+            if hasattr(relation.destination, 'crud_view') and hasattr(relation.origin, 'crud_view')
+        }
+
     class Meta(FeatureSerializer.Meta):
         exclude = ('source', 'target', 'layer', 'geom')
         fields = None